requirements.txt
atomicwrites==1.1.5
attrs==18.1.0
cachetools==2.1.0
google-api-python-client==1.7.4
google-auth==1.5.0
google-auth-httplib2==0.0.3
httplib2==0.11.3
more-itertools==4.2.0
pluggy==0.6.0
py==1.5.4
pyasn1==0.4.4
pyasn1-modules==0.2.2
pytest==3.6.3
rsa==3.4.2
six==1.11.0
uritemplate==3.0.0
